+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- added custom_plugin_dir to .griffonrc which defines a custom directory for user plugins
 
 ## [0.1.6] - 2023-03-23
 ### Changed